원래 있던 get 요청 관련 훅들을 삭제하고, SWR 라이브러리를 이용해서 다시 만듦
